Move strategy shared functions to shared
