+++ src/strategies/mod.rs
@@ -1,3 +1,4 @@
 pub mod priority_strategy;
+pub mod shared;
 pub mod types;
 pub mod uniswapx_strategy;
